Move pkg/chain into internal/pkg/base
