+++ internal/pkg/base/checks/wildcard/wildcard.go
