+++ internal/pkg/base/checks/wildcard/wildcard_test.go
